Consolidated claim's case fields into ClaimDetailsViewField where multiple properties are used across multiple area of laws
