+++ src/main/java/uk/gov/justice/laa/amend/claim/viewmodels/viewfield/ClaimDetailsViewField.java
@@ -3,6 +3,7 @@ package uk.gov.justice.laa.amend.claim.viewmodels.viewfield;
 import java.util.List;
 import java.util.function.Function;
 import lombok.Getter;
+import uk.gov.justice.laa.amend.claim.models.Claim;
 import uk.gov.justice.laa.amend.claim.models.ClaimDetails;
 
 @Getter
@@ -13,7 +14,14 @@ public enum ClaimDetailsViewField implements ClaimViewField<ClaimDetails> {
   SURNAME(new Accessor<>(ClaimDetails::getClientSurname)),
   GENDER(new Accessor<>(ClaimDetails::getClientGender), FieldOptions.GENDER),
   ETHNICITY(new Accessor<>(ClaimDetails::getClientEthnicity), FieldOptions.ETHNICITY_CODE),
-  DISABILITY(new Accessor<>(ClaimDetails::getClientDisability), FieldOptions.DISABILITY_CODE);
+  DISABILITY(new Accessor<>(ClaimDetails::getClientDisability), FieldOptions.DISABILITY_CODE),
+
+  // Common case details fields
+  CASE_REFERENCE_NUMBER(new Accessor<>(Claim::getCaseReferenceNumber)),
+  CASE_START_DATE(new Accessor<>(Claim::getCaseStartDate), FieldType.DATE),
+  UNIQUE_FILE_NUMBER(new Accessor<>(Claim::getUniqueFileNumber)),
+  STAGE_REACHED(new Accessor<>(ClaimDetails::getStageReached), FieldOptions.STAGE_REACHED),
+  CASE_CONCLUDED_DATE(new Accessor<>(Claim::getCaseEndDate), FieldType.DATE);
 
   private final Accessor<?> accessor;
   private final FieldType type;
